[DX12] Fix build issues

+++ DirectX/layers/portability/portabilityLayer.cpp
@@ -336,11 +336,13 @@ void PortabilityLayer::Post(ID3D12Device10CreatePlacedResource2Command& c) {
 void PortabilityLayer::Post(ID3D12Device5GetRaytracingAccelerationStructurePrebuildInfoCommand& c) {
   c.m_pInfo.Value->ResultDataMaxSizeInBytes *= m_AccelerationStructurePadding;
   c.m_pInfo.Value->ScratchDataSizeInBytes =
-      std::max(m_AccelerationStructureScratchMinSizeInBytes,
-               c.m_pInfo.Value->ScratchDataSizeInBytes * m_AccelerationStructureScratchPadding);
-  c.m_pInfo.Value->UpdateScratchDataSizeInBytes = std::max(
-      m_AccelerationStructureScratchMinSizeInBytes,
-      c.m_pInfo.Value->UpdateScratchDataSizeInBytes * m_AccelerationStructureScratchPadding);
+      std::max(static_cast<UINT64>(m_AccelerationStructureScratchMinSizeInBytes),
+               static_cast<UINT64>(c.m_pInfo.Value->ScratchDataSizeInBytes *
+                                   m_AccelerationStructureScratchPadding));
+  c.m_pInfo.Value->UpdateScratchDataSizeInBytes =
+      std::max(static_cast<UINT64>(m_AccelerationStructureScratchMinSizeInBytes),
+               static_cast<UINT64>(c.m_pInfo.Value->UpdateScratchDataSizeInBytes *
+                                   m_AccelerationStructureScratchPadding));
 }
 
 void PortabilityLayer::Post(
